+++ usbhub.cpp
@@ -232,9 +232,9 @@ uint8_t USBHub::Poll() {
         if(!bPollEnable)
                 return 0;
 
-        if(((long)(millis() - qNextPollTime) >= 0L)) {
+        if(((int32_t)((uint32_t)millis() - qNextPollTime) >= 0L)) {
                 rcode = CheckHubStatus();
-                qNextPollTime = millis() + 100;
+                qNextPollTime = (uint32_t)millis() + 100;
         }
         return rcode;
 }
